\name{dmr.hierclus-package}
\alias{dmr.hierclus-package}
\alias{dmr.hierclus}

\docType{package}

\title{A package with example functions from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{A package with example functions from Chapter 13 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\details{See examples in Chapter 13.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\keyword{package}
\keyword{models}

\seealso{
\code{\link[dmr.dissim:dmr.dissim-package]{dmr.dissim}}
\code{\link[dmr.kcenters:dmr.kcenters-package]{dmr.kcenters}}
}

\examples{
}
